Fix shapes translated not having correct center of mass and inertia (#133)
In Godot center of mass, inertia and mass is per the whole body, not per
shape.
In Rapier it is per shape. Instead calculate the shape's mass and
inertia in Rapier and multiply such that the whole mass equals the one
of the body.
Limitations:
Doesn't support center of body change, since we work with center of
shapes.