Fix area bug crash (#146)
Crash happens because of the use of .to() in code. Should probably
remove all these with try_to() soon. Though in this case it's good it
crashed because i found out I was converting to the wrong type. And this
value should anyway always be of type Vector2(for gravity_vector)