Implement basic sliding joint 3D (#311)
I filled out the boilerplate code that was already there. For now only
the upper and lower limits work. I am unsure on how to implement those.
However this at least provides working joints.

Another question: Is there a way to get the position of the joint Node
so we can use that orientation for the axis instead of creating it based
on the position of the physics bodies?