Stationary shapecast detection fix (#395)
Fixes a bug where shapecasts were never returning 'true', as the
'collided' state of the returned collision result wasn't being set.

NOTE: I've noticed while debugging this that the normals returned by the
shapecast seem to be backwards (eg (0,1) when they should be (0,-1)).
However, I think this might be on the Parry side...