State Management Refactor (#457)
Previous import/export functionalities for state have been removed, and
replaced by a StateManager node. This node handles state for entire
physics spaces, with no option to export or import only individual
objects.

State manager supports:
- Exporting Rapier state to a Godot Variant (can be JSON string or
binary)
- Importing Rapier state from this same exported Variant 
- Caching current Rapier state (stores physics state in memory for
convenient reloading). Cached states can have attached Godot Variant
tags, to allow users to conveniently signpost their states-- eg, they
could tag a state with a timestamp, or a tag which is a dictionary, or
just an integer corresponding to the physics frame. This is purely up to
the user. The cache also has a maximum allowed number of stored states,
which is configurable, and optional rolling functionality to
automatically remove old states if max cache length would otherwise be
exceeded.
- Three formats; 
1. None, which exports state as a plaintext JSON string (mostly for
debugging).
2. GodotBase64, which exports a string that can be conveniently used or
examined in GDScript.
3. RustBincode, which exports raw binary (fastest option).

This implementation should be more convenient for users, but still has
drawbacks (eg the scene tree can still easily be desynchronised from the
physics state when nodes are added or deleted). Future works will aim to
address this inasmuch as it is possible to do so.

As an additional note, this new state management fixes bugs which
involved areas never emitting intersection events if state was reloaded
in certain conditions.